refactor: 알림 도메인 디렉터리 구조 변경
