--- src/main/java/org/devkor/apu/saerok_server/domain/notification/application/deeplink/DeepLinkResolver.java
@@ -1,7 +1,0 @@
-package org.devkor.apu.saerok_server.domain.notification.application.deeplink;
-
-import org.devkor.apu.saerok_server.domain.notification.application.dsl.Target;
-
-public interface DeepLinkResolver {
-    String resolve(Target target);
-}
