--- src/main/java/org/devkor/apu/saerok_server/domain/notification/application/dsl/ActionKind.java
@@ -1,5 +1,0 @@
-package org.devkor.apu.saerok_server.domain.notification.application.dsl;
-
-public enum ActionKind {
-    LIKE, COMMENT, BIRD_ID_SUGGESTION
-}
