--- src/main/java/org/devkor/apu/saerok_server/domain/notification/application/dsl/TargetType.java
@@ -1,5 +1,0 @@
-package org.devkor.apu.saerok_server.domain.notification.application.dsl;
-
-public enum TargetType {
-    COLLECTION,
-}
