+++ src/main/java/org/devkor/apu/saerok_server/domain/notification/application/model/dsl/ActionKind.java
@@ -1,0 +1,5 @@
+package org.devkor.apu.saerok_server.domain.notification.application.model.dsl;
+
+public enum ActionKind {
+    LIKE, COMMENT, BIRD_ID_SUGGESTION
+}
