+++ src/main/java/org/devkor/apu/saerok_server/domain/notification/application/model/dsl/Actor.java
@@ -1,4 +1,4 @@
-package org.devkor.apu.saerok_server.domain.notification.application.dsl;
+package org.devkor.apu.saerok_server.domain.notification.application.model.dsl;
 
 public record Actor(Long id, String name) {
     public static Actor of(Long id, String name) { return new Actor(id, name); }
