+++ src/main/java/org/devkor/apu/saerok_server/domain/notification/application/model/dsl/Target.java
@@ -1,4 +1,4 @@
-package org.devkor.apu.saerok_server.domain.notification.application.dsl;
+package org.devkor.apu.saerok_server.domain.notification.application.model.dsl;
 
 public record Target(TargetType type, Long id) {
     public static Target collection(Long id) { return new Target(TargetType.COLLECTION, id); }
