--- src/main/java/org/devkor/apu/saerok_server/domain/notification/application/render/NotificationRenderer.java
@@ -1,9 +1,0 @@
-package org.devkor.apu.saerok_server.domain.notification.application.render;
-
-import org.devkor.apu.saerok_server.domain.notification.application.payload.NotificationPayload;
-
-public interface NotificationRenderer {
-    RenderedNotification render(NotificationPayload payload);
-
-    record RenderedNotification(String title, String body) { }
-}
